# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"PlaceBooksRequest": {
				"properties": {
					"bookId": {
						"type": "string"
					},
					"numberOfBooks": {
						"type": "number",
						"format": "double"
					},
					"shelf": {
						"type": "string"
					}
				},
				"required": [
					"bookId",
					"numberOfBooks",
					"shelf"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ShelfStockDto": {
				"properties": {
					"shelf": {
						"type": "string"
					},
					"count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"shelf",
					"count"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.number_": {
				"properties": {},
				"additionalProperties": {
					"type": "number",
					"format": "double"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"OrderDto": {
				"properties": {
					"orderId": {
						"type": "string"
					},
					"books": {
						"$ref": "#/components/schemas/Record_string.number_"
					},
					"fulfilled": {
						"type": "boolean"
					}
				},
				"required": [
					"orderId",
					"books"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateOrderRequest": {
				"properties": {
					"books": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"books"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FulfillmentItem": {
				"properties": {
					"book": {
						"type": "string"
					},
					"shelf": {
						"type": "string"
					},
					"numberOfBooks": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"book",
					"shelf",
					"numberOfBooks"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BookDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"author": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"image": {
						"type": "string"
					},
					"stock": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"author",
					"description",
					"price",
					"image"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FilterInput": {
				"properties": {
					"from": {
						"type": "number",
						"format": "double"
					},
					"to": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"author": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "assignment-1",
		"version": "1.0.0",
		"description": "This is a template repository for the first assignment in BVD 103 at McMaster Continuing Education.",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/warehouse/shelves": {
			"post": {
				"operationId": "PlaceBooks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Warehouse"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PlaceBooksRequest"
							}
						}
					}
				}
			}
		},
		"/warehouse/books/{bookId}/locations": {
			"get": {
				"operationId": "Locations",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ShelfStockDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Warehouse"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "bookId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/warehouse/books/{bookId}/stock": {
			"get": {
				"operationId": "Stock",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"stock": {
											"type": "number",
											"format": "double"
										},
										"bookId": {
											"type": "string"
										}
									},
									"required": [
										"stock",
										"bookId"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Warehouse"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "bookId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/orders": {
			"get": {
				"operationId": "ListOrders",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/OrderDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Orders"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"orderId": {
											"type": "string"
										}
									},
									"required": [
										"orderId"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Orders"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateOrderRequest"
							}
						}
					}
				}
			}
		},
		"/orders/{orderId}": {
			"get": {
				"operationId": "GetOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/OrderDto"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Orders"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/orders/{orderId}/fulfill": {
			"post": {
				"operationId": "FulfillOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Orders"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"fulfillment": {
										"items": {
											"$ref": "#/components/schemas/FulfillmentItem"
										},
										"type": "array"
									}
								},
								"required": [
									"fulfillment"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/books": {
			"get": {
				"operationId": "ListBooks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/BookDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Books"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "from",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "to",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "author",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateOrUpdateBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "string"
										}
									},
									"required": [
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Books"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BookDto"
							}
						}
					}
				}
			}
		},
		"/books/{id}": {
			"get": {
				"operationId": "GetBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/BookDto"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Books"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteBook",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Books"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}